+++ tests/test_console.py
@@ -119,3 +119,14 @@ def test_console_fortran_buffer():
         order="F",
         buffer=np.zeros((1, 2), order="F", dtype=tcod.console.Console.DTYPE),
     )
+
+
+def test_console_clear():
+    console = tcod.console.Console(
+        1, 1, default_fg=(1, 2, 3), default_bg=(4, 5, 6)
+    )
+    assert console.fg[0, 0].tolist() == [1, 2, 3]
+    assert console.bg[0, 0].tolist() == [4, 5, 6]
+    console.clear(fg=(7, 8, 9), bg=(10, 11, 12))
+    assert console.fg[0, 0].tolist() == [7, 8, 9]
+    assert console.bg[0, 0].tolist() == [10, 11, 12]